+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Fix (pinned) item order
+- Move prop logic into `GridImage`
 
 ## 2024-05-2X
 - Add app install button